packages/python/observability/pyproject.toml
[project]
name = "observability"
version = "0.1.0"
description = "Structured logging, OpenTelemetry tracing/metrics, and FastAPI integration for Paperwork backends"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Paperwork Labs" }]
dependencies = [
  "fastapi>=0.110,<1.0",
  "opentelemetry-api>=1.27.0",
  "opentelemetry-exporter-otlp-proto-http>=1.27.0",
  "opentelemetry-instrumentation-fastapi>=0.48b0",
  "opentelemetry-sdk>=1.27.0",
  "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
test = [
  "httpx>=0.27",
  "mypy>=1.8",
  "pytest>=8",
  "pytest-asyncio>=0.21",
  "pytest-cov>=4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/observability"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "DTZ", "SIM", "RUF", "ARG", "PIE", "C4", "T20"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
  "ARG001",
  "ARG002",
  "B008",
  "S101",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = "--import-mode=importlib --cov=observability --cov-report=term-missing --cov-fail-under=80"

[tool.coverage.run]
source_pkgs = ["observability"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
